use bigdecimal for money

+++ src/jvmMain/kotlin/model/data/Product.kt
@@ -1,8 +1,9 @@
 package model.data
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
-data class Price(@SerializedName("Betrag") val amount: Double)
+data class Price(@SerializedName("Betrag") val amount: BigDecimal)
 data class Product(
     @SerializedName("ProductId") val productId: Long,
     @SerializedName("AllergenIds") val allergenIds: List<String>,
